requirements.txt
websocket-client==1.5.1
websockets==9.1
python-dotenv==0.19.1